init.goやっぱいらん
